Moved everything to explore tool
